common+vm+builtins: make Option::Some 0 alloc
This works by replacing the previous logic. Which used a nested union
member of Value containing

    struct Option {
        .is_some=bool;
        const struct Value *value;
    }.

Thus requiring an allocation so the lifetime of the inner value is valid
throughout the program execution. The new change removes this union
member, places .is_some at the root of Value and defines Option::None as
Value.type=V_NONE, while Option::Some(<type>) is defined as
Value.type=<type> and Value.is_some=true. This allows for making
wrapping values into Option a copy, but not an allocation.