everything: started ripping the bump allocator out
This is necessary because currently a hello world is like 140MB and thats laughable shitty, the new strat is a combination of a resizable bump allocator for the compilation pipeline and a copy from space to space garbage collector like in the first lisp implementations. This is the first commit of that attempt. In optimal think a hello world would use around 400MB in pg, since:

(@println "Hello World")
1. The input is mmaped => we dont care
2. Tokens () and EOF are interned => 3 Tokens = 32*3B=96B
3. Str "Hello World" is a wrapper for a reference into the mmaped input of size 24B
4. The str has to be wrapped by a Value in the runtime: 40B
5. The bytecode has to be kept somewhere: Around 80B

Meaning: somewhere between 216B and 400B, which is like 0.00001% of the current usage. Thus the goal of this attempt is to move the real runtime memory usage towards that optimal position without completly destroying runtime performance due to many heavy reallocations