gc: reduce sizeof(GcHeader) by 8 byte by removing payload ptr
This change can be made, since GcHeader::payload is equivalent to (uint8_t)header+sizeof(GcHeader). This is the case, since each allocation passed to the garbage collector is made by the garbage collector and said allocations are prefixed with the GcHeader structure in memory. Therefore computing both the header and the payload can be done by knowing either start and either size.